Add tests from Jorgen

+++ cpp/dolfinx/mesh/utils.cpp
@@ -200,7 +200,7 @@ mesh::Mesh<T> mesh::interpolate_geometry(
       mesh::cell_type_to_basix_type(new_cmap.cell_shape()), new_cmap.degree(),
       new_cmap.variant(), basix::element::dpc_variant::unset, false);
   auto element = std::make_shared<const fem::FiniteElement<T>>(
-      b_element, std::vector<std::size_t>{gdim});
+      b_element, std::vector<std::size_t>{static_cast<std::size_t>(gdim)});
 
   fem::FunctionSpace<T> V
       = fem::create_functionspace(mesh, element, reorder_fn);
